REF: Move definitions
